Re-worked the app menu strip
Remade the menu strip as demonstrated in COMP123-2019-Week11_Lesson11. The help is a drop-down option that includes another option to open a form that shows the BMI Scale. The form is no longer an About Form. The purpose of the renamed form, "BMIScale.cs," is to simply show the BMI Scale.